+++ src/features/messages/routers/messagesRouter.ts
@@ -51,7 +51,6 @@ import {
 } from "../../../utils/error";
 import { sendFileFromRootPath } from "../../../utils/file";
 import { NotificationAttachmentDownloadMetadataResponse } from "../../../../generated/definitions/pn/NotificationAttachmentDownloadMetadataResponse";
-import { generateCheckQRPath } from "../../pn/routers/aarRouter";
 
 export const messageRouter = Router();
 const configResponse = ioDevServerConfig.messages.response;
@@ -457,53 +456,6 @@ addHandler(
     } else {
       const thirdPartyMessagePreconditions = getThirdPartyMessagePrecondition();
       res.status(200).json(thirdPartyMessagePreconditions);
-    }
-  }),
-  () => Math.random() * 500
-);
-
-addHandler(
-  messageRouter,
-  "post",
-  addApiV1Prefix("/send/aar"),
-  lollipopMiddleware(async (req, res) => {
-    const sendQRCodeUrl = `${serverUrl}${generateCheckQRPath()}`;
-    const sendQRCodeBody = JSON.stringify(req.body);
-    try {
-      const sendQRCodeResponse = await fetch(sendQRCodeUrl, {
-        method: "post",
-        headers: {
-          ...MessagesService.lollipopClientHeadersFromHeaders(req.headers),
-          ...MessagesService.generateFakeLollipopServerHeaders(
-            ioDevServerConfig.profile.attrs.fiscal_code
-          ),
-          ...MessagesService.sendAPIKeyHeader(),
-          ...MessagesService.sendTaxIdHeader(
-            ioDevServerConfig.profile.attrs.fiscal_code
-          ),
-          "Content-Type": "application/json"
-        },
-        body: sendQRCodeBody
-      });
-
-      const contentType = sendQRCodeResponse.headers.get("content-type");
-      const responseBodyBuffer = await sendQRCodeResponse.arrayBuffer();
-      const body = Buffer.from(responseBodyBuffer);
-      if (contentType) {
-        res.setHeader("Content-Type", contentType);
-      }
-      res.status(sendQRCodeResponse.status).send(body);
-    } catch (e) {
-      const errorMessage = unknownToString(e);
-      res
-        .status(500)
-        .json(
-          getProblemJson(
-            500,
-            "QRCode unexpected error",
-            `Unexpected error while contacting SEND QRCode endpoint (${errorMessage})`
-          )
-        );
     }
   }),
   () => Math.random() * 500
